+++ dlib/math/utils.d
@@ -208,41 +208,6 @@ unittest
 }
 
 /**
- * Sum of all elements of an array
- */
-deprecated("use reduce!((a, b) => a + b) instead")
-T sum(T) (T[] array...) nothrow
-{
-    T result = 0;
-    foreach(v; array)
-        result += v;
-    return result;
-}
-
-/**
- * Negate all elements of an array
- */
-deprecated("use map!(a => -a) instead")
-T[] invertArray(T) (T[] array...) nothrow
-{
-    auto result = new T[array.length];
-    foreach(i, v; array)
-        result[i] = -v;
-    return result;
-}
-
-/**
- * If all elements are zeros
- */
-deprecated("use reduce!((a, b) => a + b == 0) instead")
-bool allIsZero(T) (T[] array...) nothrow
-{
-    foreach(i, v; array)
-        if (v != 0) return false;
-    return true;
-}
-
-/**
  * If at least one element is zero
  */
 bool oneOfIsZero(T) (T[] array...) nothrow